Migrate Opensearch to IndexerInterface and SearcherInterface
