+++ packages/seal-opensearch-adapter/OpensearchAdapter.php
@@ -4,22 +4,27 @@ namespace Schranz\Search\SEAL\Adapter\Opensearch;
 
 use OpenSearch\Client;
 use Schranz\Search\SEAL\Adapter\AdapterInterface;
-use Schranz\Search\SEAL\Adapter\ConnectionInterface;
+use Schranz\Search\SEAL\Adapter\IndexerInterface;
 use Schranz\Search\SEAL\Adapter\SchemaManagerInterface;
+use Schranz\Search\SEAL\Adapter\SearcherInterface;
 
 final class OpensearchAdapter implements AdapterInterface
 {
-    private readonly ConnectionInterface $connection;
-
     private readonly SchemaManagerInterface $schemaManager;
+
+    private readonly IndexerInterface $indexer;
+
+    private readonly SearcherInterface $searcher;
 
     public function __construct(
         private readonly Client $client,
-        ?ConnectionInterface $connection = null,
         ?SchemaManagerInterface $schemaManager = null,
+        ?IndexerInterface $indexer = null,
+        ?SearcherInterface $searcher = null,
     ) {
-        $this->connection = $connection ?? new OpensearchConnection($client);
         $this->schemaManager = $schemaManager ?? new OpensearchSchemaManager($client);
+        $this->indexer = $indexer ?? new OpensearchIndexer($client);
+        $this->searcher = $searcher ?? new OpensearchSearcher($client);
     }
 
     public function getSchemaManager(): SchemaManagerInterface
@@ -27,8 +32,13 @@ final class OpensearchAdapter implements AdapterInterface
         return $this->schemaManager;
     }
 
-    public function getConnection(): ConnectionInterface
+    public function getIndexer(): IndexerInterface
     {
-        return $this->connection;
+        return $this->indexer;
+    }
+
+    public function getSearcher(): SearcherInterface
+    {
+        return $this->searcher;
     }
 }
