+++ packages/seal-opensearch-adapter/Tests/OpensearchIndexerTest.php
@@ -3,9 +3,9 @@
 namespace Schranz\Search\SEAL\Adapter\Opensearch\Tests;
 
 use Schranz\Search\SEAL\Adapter\Opensearch\OpensearchAdapter;
-use Schranz\Search\SEAL\Testing\AbstractAdapterTestCase;
+use Schranz\Search\SEAL\Testing\AbstractIndexerTestCase;
 
-class OpensearchAdapterTest extends AbstractAdapterTestCase
+class OpensearcnIndexerTest extends AbstractIndexerTestCase
 {
     public static function setUpBeforeClass(): void
     {
